Refactor expression parser.
